Fix to user position harvesting
